move labels to Type namespace
